requirements.txt
pydantic==2.11.4
pydantic-ai==0.2.3
python-dotenv==1.1.0
